move getTextStreamData to FileUtil
